Add missing return statement in setValue
All automatically generated setter methods return "this" to allow method chaining. This is true of any given SAPUI5 control and is automatically done by ManagedObject. However, this custom control implementation does not comply with this behavior.

This commit adds the missing return statement in setValue to comply with the behavior of other SAPUI5 controls.