Update step-23-filtering-5295470.md
Removing the formatter property in the controller.
The formatter property is not used (even from the associated view) and is assigned a nonexistent value. 
It seems that it is a leftover from an old version which no longer has any reason to exist; now
formatter modules are directly loaded into the view via sap.ui.core.require.